# fix the data dependent initializaiton.