See issue https://github.com/openai/pixel-cnn/issues/36
The output of conv2d should be updated after g and b are updated in data dependent initialization.

The loss decreases faster  when using the new data dependent initialization on cifar-10 dataset. I only tested the first few epochs for the time being. I will update the running log later.

new initialization:
Iteration 0, time = 3317s, train bits_per_dim = 4.1773, test bits_per_dim = 7.6512                                                     
Iteration 1, time = 3275s, train bits_per_dim = 3.7190, test bits_per_dim = 4.4990

original initialization:
Iteration 0, time = 3317s, train bits_per_dim = 4.1814, test bits_per_dim = 11.5115                                                    
Iteration 1, time = 3277s, train bits_per_dim = 3.6735, test bits_per_dim = 8.9926 
Iteration 2, time = 3277s, train bits_per_dim = 3.5277, test bits_per_dim = 5.1907